backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-RESTful
requests
psycopg2
python-dotenv
pytest